// apps/api/src/resources/credits/entities/credits.enums.ts
export enum CreditsPlatform {
  MCOM_LOYALTY = "MCOM_LOYALTY",
  MCOM_MALL = "MCOM_MALL",
}

export enum CreditsRewardType {
  PERCENTAGE = "PERCENTAGE",
  FIXED = "FIXED",
}

export enum CreditsTransactionType {
  CREDIT = "CREDIT",
  DEBIT = "DEBIT",
}

export enum CreditsUnit {
  CREDITS = "CREDITS",
  GBP = "GBP",
}

export enum CreditsUserType {
  PARTICIPANT = "PARTICIPANT",
  BUSINESS = "BUSINESS",
}